prototype/requirements.txt
streamlit==1.55.0
llama-index==0.14.19
openai==2.25.0
faiss-cpu==1.13.2
PyMuPDF==1.23.0
python-pptx==1.0.2
paddleocr==3.4.0
openpyxl==3.1.5
pandas==2.3.3
numpy==2.4.2
sentence-transformers
torch
transformers